requirements.txt

certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.1
cx-Oracle==8.3.0
cycler==0.12.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
greenlet==3.0.3
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
mysql-connector==2.2.9
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pip==24.0
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.3
schedule==1.2.1
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
shap==0.45.0
six==1.16.0
slicer==0.0.7
SQLAlchemy==2.0.29
sympy==1.12
threadpoolctl==3.4.0
timm==0.9.16
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
